# Use custom tokenization function for NetCDFDataProxy objects