## 🚀 Pull Request

### Description
<!-- Provide a clear description about your awesome pull request -->
<!-- Tell us all about your new feature, improvement, or bug fix -->


Use custom tokenization function for NetCDFDataProxy objects. This speeds up loading small NetCDF files. With the multiple_cubes.nc file from https://github.com/SciTools/iris/issues/6223, the speedup is 10 to 20 %.

---
[Consult Iris pull request check list]( https://scitools-iris.readthedocs.io/en/latest/developers_guide/contributing_pull_request_checklist.html)

---
Add any of the below labels to trigger actions on this PR:

- https://github.com/SciTools/iris/labels/benchmark_this