@pp-mo 2024-12-11
> the speedup is 10 to 20 %, though in the benchmarks below it is not really significant.

Looks really neat + I don't doubt you're right this is worth it for some cases.
I just wonder though, is it worth _adding_ a benchmark which demonstrates the advantage ?