@bouweandela 2024-12-11
The benchmark I added just shows:
```
| Change   | Before [0fdedb46]    | After [23e9462b]    |   Ratio | Benchmark (Parameter)                                                                       |
|----------|----------------------|---------------------|---------|---------------------------------------------------------------------------------------------|
|          | 1.10±0.01s           | 1.06±0.01s          |    0.96 | load.ManyCubes.time_many_cube_load                                                          |
```

On my laptop the timings are a bit more positive, but maybe this isn't really worth the effort. Shall I just close this?