@bouweandela 2024-12-20
As discussed at the peloton meeting, this is probably useful after all. The benchmark is likely slower in CI than on my laptop because it has different hardware and therefore spends more time on I/O, reducing the relative speedup of this change on that particular machine.