@bouweandela 2025-02-13
The improvement is still ~50 ms, but since the overall time is now shorter the benefit becomes relatively larger. CI results:
```
| Change   | Before [ba9a6e2d]    | After [a09fe715]    |   Ratio | Benchmark (Parameter)                                                                       |
|----------|----------------------|---------------------|---------|---------------------------------------------------------------------------------------------|
|          | 469±10ms             | 417±4ms             |    0.89 | load.ManyCubes.time_many_cube_load                                                          |
```
and on my laptop
```

| Change   | Before [ba9a6e2d] <main>   | After [9a1a14df] <netcdfdataproxy-dask-tokenize>   |   Ratio | Benchmark (Parameter)              |
|----------|----------------------------|----------------------------------------------------|---------|------------------------------------|
|          | 177±0.9ms                  | 154±4ms                                            |    0.87 | load.ManyCubes.time_many_cube_load |
```