@bouweandela 2025-02-14
Benchmark result from my laptop with the most recent commit (`python bm_runner.py branch upstream/main --bench=load.ManyCubes.time_many_cube_load`):

```
| Change   | Before [ba9a6e2d] <main>   | After [31abeaf7] <netcdfdataproxy-dask-tokenize>   |   Ratio | Benchmark (Parameter)              |
|----------|----------------------------|----------------------------------------------------|---------|------------------------------------|
| -        | 168±7ms                    | 102±1ms                                            |    0.61 | load.ManyCubes.time_many_cube_load |
````