@bouweandela 2025-02-14
It looks like the advantage is rather minimal for the benchmarked case now that we are creating fewer arrays since #6252. Let's just close this.