@ukmo-ccbunney 2025-03-16
> It looks like the advantage is rather minimal for the benchmarked case now that we are creating fewer arrays since #6252. Let's just close this.

Sincere apologies for not getting round to looking at this in good time @bouweandela.